voc #150141 게시판 : 공개/비공개 설정

+++ src/main/java/egovframework/ezEKP/ezBoard/web/EzBoardAdminController.java
@@ -1022,7 +1022,7 @@ public class EzBoardAdminController extends EgovFileMngUtil {
 		int tabNum = 3; //탭 개수
 		
 		/* 2023-11-08 민지수 - 카테고리게시판으로 게시판 유형을 변경한 경우, 즐겨찾기와 마이게시판 목록에서 해당 게시판 제거 */
-		if (boardpro.getGuBun().equals("10")) {
+		if (boardpro.getGuBun() != null && boardpro.getGuBun().equals("10")) {
 			ezBoardAdminService.deleteMyBoardData("MyBoards", boardPropertyVO.getBoardID(), userInfo.getTenantId());
 			ezBoardAdminService.deleteMyBoardData("MyBoardTree", boardPropertyVO.getBoardID(), userInfo.getTenantId());
 		}
